+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The by month exports do not include conversion projects with a revoked DAO.
 - The single by month view and export are now sorted by the conversion or
   transfer date only.
+- The pre conversion grants export includes a new column that shows the DAO
+  revocation state of a project as applicable.
 
 ## [Release-76][release-76]